+++ cmd/istioctl/BUILD
@@ -47,10 +47,10 @@ go_test(
     ],
     library = ":go_default_library",
     deps = [
+        "//adapter/config/tpr:go_default_library",
         "//apiserver:go_default_library",
         "//client/proxy:go_default_library",
         "//cmd/version:go_default_library",
-        "//platform/kube:go_default_library",
         "//test/util:go_default_library",
         "@com_github_ghodss_yaml//:go_default_library",
         "@com_github_googleapis_googleapis//:go_default_library",
