+++ cmd/version/BUILD
@@ -12,7 +12,7 @@ go_library(
 
 go_test(
     name = "go_default_test",
-    srcs = ["version_test.go"],
     size = "small",
+    srcs = ["version_test.go"],
     library = ":go_default_library",
 )
